feat : 액티비티에 적용
